[aprism] Contrib folder - Better cleaning Bis

+++ @prism/tools/building/02-contrib/02-CLEAN-contrib-folders/CLEAN-PROD-lua.py
@@ -12,7 +12,8 @@ BUILD_TOOLS_DIR = THIS_DIR.parent.parent
 sys.path.append(str(BUILD_TOOLS_DIR))
 
 
-from cbutils.core import *
+from cbutils.core   import *
+from cbutils.texnco import *
 
 # -- IMPORT CBUTILS - END -- #
 # -------------------------- #
@@ -24,14 +25,37 @@ from shutil import rmtree
 # -- CONSTANTS #1 -- #
 # ------------------ #
 
-GLOB_BY_DIRS = {
+COMMON_EXTS = [
+    "pdf",
+] + LATEX_TEMP_EXT
+
+FILE_GLOB_BY_DIRS = {
     'dev': [
-        "*/*.pdf",
-        "*/*.lua",
-        "*/*.sty",
+        f"*/*.{ext}"
+        for ext in [
+            "lua",
+            "sty",
+        ] + COMMON_EXTS
     ],
     'fake-prod/showcase': [
-        "*/*.pdf",
+        f"*/*.{ext}"
+        for ext in COMMON_EXTS
+    ],
+}
+
+
+COMMON_FOLDER = [
+    "_*",
+]
+
+FOLDER_GLOB_BY_DIRS = {
+    'dev': [
+        f"*/{g}"
+        for g in COMMON_FOLDER
+    ],
+    'fake-prod/showcase': [
+        f"*/{g}"
+        for g in COMMON_FOLDER
     ],
 }
 
@@ -56,9 +80,10 @@ CONTRIB_PROD_DIR = (
 
 relpath = CONTRIB_PROD_DIR.relative_to(PROJ_DIR)
 
-logging.info(f"Clean '{relpath}' folder.")
 
-for subdir, patterns in GLOB_BY_DIRS.items():
+logging.info(f"Clean 'files in {relpath}' folder.")
+
+for subdir, patterns in FILE_GLOB_BY_DIRS.items():
     subpath = CONTRIB_PROD_DIR
 
     for p in subdir.split('/'):
@@ -66,4 +91,19 @@ for subdir, patterns in GLOB_BY_DIRS.items():
 
         for g in patterns:
             for f in subpath.glob(g):
-                f.unlink()
+                if f.is_file():
+                    f.unlink()
+
+
+logging.info(f"Clean 'dirs in {relpath}' folder.")
+
+for subdir, patterns in FOLDER_GLOB_BY_DIRS.items():
+    subpath = CONTRIB_PROD_DIR
+
+    for p in subdir.split('/'):
+        subpath /= p
+
+        for g in patterns:
+            for d in subpath.glob(g):
+                if d.is_dir():
+                    rmtree(d)
